requirements.txt
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
redis==4.5.3
celery==5.3.1
django-cors-headers==3.14.0
